# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/slots": {
      "get": {
        "description": "Requiere venueId, sportId y date. Solo devuelve turnos AVAILABLE y futuros del día local en Buenos Aires. Un día sin turnos devuelve []. Incluye los turnos del horario recurrente, materializados para el día consultado de forma idempotente. Excluye cierres por sede y horarios automáticos obsoletos.",
        "operationId": "SlotsController_list",
        "parameters": [
          {
            "name": "venueId",
            "required": true,
            "in": "query",
            "description": "Sede elegida; filtro obligatorio",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "sportId",
            "required": true,
            "in": "query",
            "description": "Deporte elegido; filtro obligatorio",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Día local en America/Argentina/Buenos_Aires.",
            "schema": {
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2099-01-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SlotResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Consultar disponibilidad por fecha",
        "tags": [
          "Turnos"
        ]
      },
      "post": {
        "description": "La sede debe ofrecer el deporte. Fin posterior al inicio. No se impone duración ni política de solapamientos: una sede puede tener varias canchas.",
        "operationId": "SlotsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlotDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Crear un turno",
        "tags": [
          "Turnos"
        ]
      }
    },
    "/api/slots/{id}": {
      "get": {
        "description": "Puede devolver turnos pasados o bloqueados; esta lectura no acredita disponibilidad.",
        "operationId": "SlotsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID público del registro.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Obtener un turno",
        "tags": [
          "Turnos"
        ]
      },
      "patch": {
        "description": "Usar status UNAVAILABLE para bloquearlo. Campos omitidos se conservan; null no es válido.",
        "operationId": "SlotsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID público del registro.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSlotDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Actualizar o bloquear un turno",
        "tags": [
          "Turnos"
        ]
      },
      "delete": {
        "description": "Si hay borradores asociados devuelve 409; bloquearlo en su lugar.",
        "operationId": "SlotsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID público del registro.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado. Sin cuerpo de respuesta."
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Eliminar un turno sin solicitudes",
        "tags": [
          "Turnos"
        ]
      }
    },
...
  },
  "info": {
    "title": "API del turnero de polideportivos",
    "description": "Flujo: deporte → contacto → zona y sede → detalle → fecha y turno → vista previa de WhatsApp. Los catálogos empiezan vacíos. Cargarlos con clave de gestión en el orden deportes, zonas, sedes, relaciones sede/deporte y turnos. Los UUID de los ejemplos son ficticios. Generar WhatsApp no envía mensajes, no bloquea disponibilidad y no confirma reservas. Los borradores se acceden mediante UUID secreto, sin login. JSON de hasta 16 KiB; 100 solicitudes/minuto/IP.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "ManagementKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clave de gestión solo para operaciones protegidas. No se almacena en la documentación."
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "Código de estado HTTP."
          },
          "message": {
            "description": "Mensaje de negocio o lista de errores de validación.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "El turno no está disponible"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Nombre del error; puede omitirse, por ejemplo en el límite de solicitudes."
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "SlotResponseDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid",
            "description": "Sede habilitada que ofrece el deporte",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "sportId": {
            "type": "string",
            "format": "uuid",
            "description": "Deporte del turno",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "description": "Inicio con zona horaria explícita. Se normaliza a UTC.",
            "pattern": "T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$",
            "example": "2099-01-01T18:00:00-03:00"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fin posterior al inicio, con zona horaria explícita. Se normaliza a UTC.",
            "pattern": "T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$",
            "example": "2099-01-01T19:00:00-03:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE",
              "RESERVED"
            ],
            "description": "Estado efectivo. RESERVED solo se obtiene por confirmación administrativa.",
            "example": "AVAILABLE"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador público del turno",
            "example": "00000000-0000-4000-8000-000000000001"
          }
        },
        "required": [
          "venueId",
          "sportId",
          "startsAt",
          "endsAt",
          "status",
          "id"
        ]
      },
...
      "CreateSlotDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid",
            "description": "Sede habilitada que ofrece el deporte",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "sportId": {
            "type": "string",
            "format": "uuid",
            "description": "Deporte del turno",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "description": "Inicio con zona horaria explícita. Se normaliza a UTC.",
            "pattern": "T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$",
            "example": "2099-01-01T18:00:00-03:00"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fin posterior al inicio, con zona horaria explícita. Se normaliza a UTC.",
            "pattern": "T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$",
            "example": "2099-01-01T19:00:00-03:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "description": "Disponibilidad del turno.",
            "example": "AVAILABLE"
          }
        },
        "required": [
          "venueId",
          "sportId",
          "startsAt",
          "endsAt",
          "status"
        ],
        "additionalProperties": false
      },
      "UpdateSlotDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid",
            "description": "Sede habilitada que ofrece el deporte",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "sportId": {
            "type": "string",
            "format": "uuid",
            "description": "Deporte del turno",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "description": "Inicio con zona horaria explícita. Se normaliza a UTC.",
            "pattern": "T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$",
            "example": "2099-01-01T18:00:00-03:00"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fin posterior al inicio, con zona horaria explícita. Se normaliza a UTC.",
            "pattern": "T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{1,3})?(?:Z|[+-]\\d{2}:\\d{2})$",
            "example": "2099-01-01T19:00:00-03:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "description": "Disponibilidad del turno.",
            "example": "AVAILABLE"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}